alterando tipo de variáveis para double

+++ include/objeto.h
@@ -8,12 +8,12 @@
 typedef struct Objeto {
     int id;
     double x, y;
-    int largura;
+    double largura;
 
-    int inicioX, fimX;
+    double inicioX, fimX;
 } Objeto_t;
 
-Objeto_t CriaObj(int id, double x, double y, int largura);
+Objeto_t CriaObj(int id, double x, double y, double largura);
 Objeto_t* EncontrarObj(Objeto_t* vetor, int num_objetos, int id);
 void DestroiObj(Objeto_t o);
 void AttPos(Objeto_t *o, double novo_x, double novo_y);
